Add files via upload
So the server.py file was not made by me but it is needed to run the support_functions.py file which was created by me. I created this as my special mini-ChatGPT where users will ask questions and will get answers as long as the specific keywords are mentioned. My keywords are "skills", "experiences", "education",  and "name". Fell free to run it so you get to know more about me, Ayomide Akinsanya