Add files via upload
This is a Python program I wrote to company fertility rate between years. The user will enter as many data inputs as they want and they can can be able to see how the fertility rate change between the two years whether it was an increase decrease, or the same.