Add files via upload
The game.py file which is the background was not made by me but the board.py was made by me which includes the structures, conditions for winning and losing as well as well as the idea and logic behind the game. It is a board game where the draft taken from the map.txt file, players positions from the players.txt file, and the exit locations, from the exit.txt file all come together to creating a board game which you can control using your keyboard. The minimum players needed is 1 and the maximum is 4.